% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personsOutput.R
\name{boxplotScoreDifferences}
\alias{boxplotScoreDifferences}
\title{Boxplot of Score Differences}
\usage{
boxplotScoreDifferences(personTibble_base, personTibble_policy)
}
\arguments{
\item{personTibble_base}{persons tibble of the base case, can be loaded with readPersonsTable.}

\item{personTibble_policy}{persons tibble of the policy case, can be loaded with readPersonsTable.}
}
\value{
ggplot boxplot of the distribution of the score differences
}
\description{
The \code{boxplotScoreDifferences} function generates a boxplot to compare the score differences between two sets of data
represented by personTibble_base and personTibble_policy tibbles.
It provides insights into the distribution of score differences between the two sets of data.
}
