% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripsOutput.R
\name{compareBasePolicyOutput}
\alias{compareBasePolicyOutput}
\title{Creates dashboard for the given table or folder with data}
\usage{
compareBasePolicyOutput(
  baseFolder,
  policyFolder,
  dump.output.to = matsimDumpOutputDirectory
)
}
\arguments{
\item{dump.output.to}{folder that saves and configures yaml for simwrapper dashboard and all plots using functions:
plotModalSplitBarChart(),plotModalSplitPieChart(),plotModalShift().}

\item{folder}{specifies data source folder with tripsOutput}

\item{append}{specifies if the ouput folder should be erased before creating}
}
\value{
tibble of output_trips from folder
}
\description{
Creates dashboard for the given table or folder with data
}
