% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{compareTripTypesBarChart}
\alias{compareTripTypesBarChart}
\title{Creates BarChart of changing trip types(originating,transit etc) between 2 tables
and saves output to dump.output.to}
\usage{
compareTripTypesBarChart(
  tripsTable1,
  tripsTable2,
  shapeTable,
  crs,
  dump.output.to = matsimDumpOutputDirectory
)
}
\arguments{
\item{tripsTable1}{tibble of trips_output (from readTripsTable(),f.e. base case)}

\item{tripsTable2}{tibble of trips_output (from readTripsTable(),f.e. policy case)}

\item{shapeTable}{sf object(data.frame with geometries), can be received by using st_read(path_to_geographical_file)}

\item{crs}{numeric of EPSG code or proj4string, can be found in network file from output directory of MATSim simulation}

\item{dump.output.to}{folder that saves resulting image of BarChart}
}
\value{
plot with percentage of each type of trips between 2 tables
}
\description{
Creates BarChart of changing trip types(originating,transit etc) between 2 tables
and saves output to dump.output.to
}
