% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{createCountScatterPlot}
\alias{createCountScatterPlot}
\title{Creates a Via-Style scatterplot for each run}
\usage{
createCountScatterPlot(
  joinedFrame,
  ll = ~0.8 * -200,
  ul = ~x * 1.2 + 200,
  threshold = 100
)
}
\arguments{
\item{joinedFrame}{A tibble from mergeCountsAndLinks}

\item{ll}{Formula to calculate lower limit of the quality label 'exact', default = 0.8*x - 200}

\item{ul}{Formula to calculate lower limit of the quality label 'exact', default = 1.2*x + 200}

\item{threshold}{Threshold from which data is scaled to log10.}
}
\value{
A ggplot Scatterplotplot, which can be adjusted, if needed.
}
\description{
Takes a tibble from mergeCountsAndLinks.
A scatterplot with counts on the x axis and matsim dtv on the y axis is created and colored
by the road type.
Lower and upper Limits define the section which is considered as an 'exact' estimation. Limits
are defined by custom formulas.
}
\details{
The function calls the matsim-r function processDtvEstimationQuality which is handeling the limits.
}
