% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripsOutput.R
\name{deriveODMatrix}
\alias{deriveODMatrix}
\title{Creates an instance of ODMatrix(origin/destination) in conventional form or for the simwrapper}
\usage{
deriveODMatrix(
  tripsTable,
  shapePath,
  crs,
  dump.output.to = matsimDumpOutputDirectory,
  simwrapper = FALSE,
  colnames = "numeric"
)
}
\arguments{
\item{tripsTable}{table of output trips(from readTripsTable)}

\item{shapePath}{full path to shapefile}

\item{crs}{numeric of EPSG code or proj4string, can be found in network file from output directory of MATSim simulation}

\item{dump.output.to}{path to a folder to save csv file of ODMatrix}

\item{simwrapper}{create output in a simwrapper form if set to path of the shapefile}

\item{colnames}{if the specific shapefile contains known columns, they could be specified as name for columns OD}
}
\value{
tibble of origin/destination matrix
}
\description{
Creates an instance of ODMatrix(origin/destination) in conventional form or for the simwrapper
}
