% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripsOutput.R
\name{filterByRegion}
\alias{filterByRegion}
\title{Filtering of trips_table(from readTripsTable) depending on how they located in given shape}
\usage{
filterByRegion(
  tripsTable,
  shapeTable,
  crs,
  start.inshape = TRUE,
  end.inshape = TRUE
)
}
\arguments{
\item{shapeTable}{sf object(data.frame with geometries), can be received by using st_read(path_to_geographical_file)}

\item{crs}{numeric of EPSG code or proj4string, can be found in network file from output directory of MATSim simulation}

\item{start.inshape}{bool, defines trips to conclude (see Description)}

\item{end.inshape}{bool, defines trips to conclude (see Description)}

\item{table}{tibble of trips_output (from readTripsTable())}
}
\value{
tibble, with filtered trips depending on shapeTable and special flags (see Description)
}
\description{
Takes trips_table and shapeTable(sf object from file representing geographical data, can be received by using function st_read(path_to_file))
transforms both objects to match mutual CRS(network.xml from MATSimOutputDirectory)
and filters the trips from table depending on *.inshape flags:
if start.inshape = TRUE & end.inshape = TRUE return table that contains trips inside given shape
if start.inshape = TRUE & end.inshape = FALSE return table that contains trips which starts in shape and ends out of the shape
if start.inshape = FALSE & end.inshape = TRUE return table that contains trips which ends in shape and starts out of the shape
if start.inshape = FALSE & end.inshape = FALSE return table that contains trips which starts and ends our of the given shape
}
