% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simwrapper.R
\name{generateXYHexagonYaml}
\alias{generateXYHexagonYaml}
\title{Creates/adds XY hexagon definition of output trips to a summary dashboard}
\usage{
generateXYHexagonYaml(
  pathToOutputTrips,
  crs,
  dump.output.to = matsimDumpOutputDirectory
)
}
\arguments{
\item{pathToOutputTrips}{file path to output_trips.csv file}

\item{crs}{specifies the coordinate reference system of output_trips}

\item{dump.output.to}{path to a folder with simwrapper dashboard}
}
\value{
changed file
}
\description{
Creates/adds XY hexagon definition of output trips to a summary dashboard
}
