% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{getCrsFromConfig}
\alias{getCrsFromConfig}
\title{Reads an coordinate referenec system of MATSim output directory
from output_config.xml}
\usage{
getCrsFromConfig(folder)
}
\arguments{
\item{folder}{specifies path to find config}
}
\value{
code of coordinate reference system
}
\description{
Reads an coordinate referenec system of MATSim output directory
from output_config.xml
}
