% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personsOutput.R
\name{join_base_and_policy_persons}
\alias{join_base_and_policy_persons}
\title{Joins base and policy person tables}
\usage{
join_base_and_policy_persons(base_persons, policy_persons, crs = "EPSG:25832")
}
\arguments{
\item{base_persons}{persons tibble of the base case, can be loaded with readPersonsTable.}

\item{policy_persons}{persons tibble of the policy case, can be loaded with readPersonsTable.}

\item{crs}{coordinate reference system for first_act_x and first_act_y coordinates; enter in form of EPSG code. Default is EPSG:25832}
}
\value{
tibble of joined persons
}
\description{
Uses inner_join to combine base and policy persons tables.
Calculates difference in score between base and policy cases.
}
