% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{loadNetwork}
\alias{loadNetwork}
\title{Load MATSim network into memory}
\usage{
loadNetwork(filename)
}
\arguments{
\item{filename}{File to load. Can be XML or gzipped XML}
}
\value{
"nodes" and "links" tibbles in a list object.
}
\description{
Loads a MATSim XML network file, merging node, link, and attribute elements
into a nodes tibble and a consolidated links tibble which contains
joined columns from attributes and nodes such as coordinates.
}
