% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{mergeCountsAndLinks}
\alias{mergeCountsAndLinks}
\title{Load Counts, a limited number of Linkstats and Network links as joined tibble into memory}
\usage{
mergeCountsAndLinks(
  counts,
  network,
  linkStatsList,
  sampleSize = 0.25,
  networkModes = c("car")
)
}
\arguments{
\item{counts}{Tibble with counts data}

\item{network}{Tibble with network nodes and links}

\item{linkStatsList}{List with filepaths to Linkstats, uses the runId as key, and filepath as value}

\item{sampleSize}{sample size of the scenario between 0 and 1 to rescale DTV values}

\item{networkModes}{a vector with network modes, which are needed for analysis}
}
\value{
Tibble with MATSim link id as key ("loc_id"), DTV from MATSim runs and link type
}
\description{
Function to import and join MATSim Counts, Linkstats and network link types
Linkstats will be aggregated to DTV. An additional sampleSize parameter can be
used to scale DTV values.
The parameter networkModes contains a set of network modes to filter and select
columns of link stats.
}
\examples{

runList = list("run_1" = "path/to/linkstats_1",
              "run_2" = "path/to/linkstats_2"
              )

mergeCountsAndLinks(
 counts = countsDataFrame,
 network = networkDataFrame,
 linkStatsList = runList,
 sampleSize = 0.25,
 networkModes = c("car", "bike", "freight")
 )


}
