% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personsOutput.R
\name{persons_attributes_on_hex_grid}
\alias{persons_attributes_on_hex_grid}
\title{Overlays hexagonal grid over study area, and creates summary statistics for each pixel}
\usage{
persons_attributes_on_hex_grid(persons, shp, n = 20)
}
\arguments{
\item{persons}{simple feature dataframe containing persons, including score_base, score_policy, and score_diff, income, carAvail, sim_ptAbo}

\item{shp}{shape file onto which grid is overlaid}

\item{n}{integer of length 1 or 2, number of grid cells in x and y direction (columns, rows)}
}
\value{
sf object of hexagonal grid
}
\description{
CRS of shape file must match CRS of persons sf object
}
