% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{plot_arrtime_by_act}
\alias{plot_arrtime_by_act}
\title{Create a line plot of arrival time by activity}
\usage{
plot_arrtime_by_act(
  trips_table,
  unite_activities = character(0),
  united_name = "united"
)
}
\arguments{
\item{trips_table}{tibble of trips_output (from \link{read_output_trips})}

\item{unite_activities}{optional, vector of character strings,
changes names of chosen categories to a new specified name (i.e. school and university to education)}

\item{united_name}{optional, character string, specifies the new name of the united activities}
}
\value{
A plotly object representing the line plot of arrival time by activity.
}
\description{
Takes the tibble created by \link{read_output_trips}
to create a line plot of the
count of destination activities for a specific arrival time
Using the parameter unite_activities, specific activities can be renamed into one with the name specified with united_name (by default 'united')
}
