% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{plot_compare_count_by_spatialcat}
\alias{plot_compare_count_by_spatialcat}
\title{Bar Chart with main_mode on x-axis and average travel/wait time on y-axis
XXXX
Takes Table trips_output (from readTripsTable()),
to plot bar chart with with values that represent
time spent on traveling/waiting
Using parameters unite.columns, specific columns could be given, to unite them in 1 mode with the name united.name(by default 'united')}
\usage{
plot_compare_count_by_spatialcat(
  trips_table1,
  trips_table2,
  shape_table,
  crs,
  dump.output.to = matsimDumpOutputDirectory
)
}
\arguments{
\item{tripsTable1}{tible of trips_output (from readTripsTable())}

\item{tripsTable2}{tible of trips_output (from readTripsTable())}

\item{unite.columns}{vector of character strings, that represent patterns of columns to be united, changes name of all transport modes in the tibble copy to united.name = "united" that matches PATTERNS given in unite.columns}

\item{united.name}{character string, if columns were united, you can specify name for the resulting column in chart}
}
\value{
Bar Chart plot of average time spent on travel/wait
}
\description{
Bar Chart with main_mode on x-axis and average travel/wait time on y-axis
XXXX
Takes Table trips_output (from readTripsTable()),
to plot bar chart with with values that represent
time spent on traveling/waiting
Using parameters unite.columns, specific columns could be given, to unite them in 1 mode with the name united.name(by default 'united')
}
