% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{plot_compare_count_by_spatialcat_barchart}
\alias{plot_compare_count_by_spatialcat_barchart}
\title{Plot comparison bar chart of count by spatial category}
\usage{
plot_compare_count_by_spatialcat_barchart(
  trips_table1,
  trips_table2,
  shape_table,
  crs
)
}
\arguments{
\item{trips_table1}{tibble of trips_output (from \link{read_output_trips})}

\item{trips_table2}{tibble of trips_output (from \link{read_output_trips})}

\item{shape_table}{sf object(data frame with geometries), can be created using st_read(), is used to categorize the trips.}

\item{crs}{numeric representation of the EPSG code or proj4string for the corresponding coordinate system of the trip coordinates,
can be found in network file from output directory of MATSim simulation}
}
\value{
Bar chart comparing the number of trips by spatial category.
}
\description{
Takes two tibbles from \link{read_output_trips} to
generate a bar chart comparing the number of trips across different spatial categories.
}
