% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{plot_compare_mainmode_barchart}
\alias{plot_compare_mainmode_barchart}
\title{Plot bar chart of changes in modal split}
\usage{
plot_compare_mainmode_barchart(
  trips_table1,
  trips_table2,
  unite.modes = character(0),
  united.name = "united"
)
}
\arguments{
\item{unite.modes}{vector of character strings,
changes names of chosen modes in the column main_mode to a new chosen name (i.e. drtNorth and drtSouth to drt),
using the function (\link{process_rename_mainmodes})}

\item{united.name}{character string, specifies the name of the united mode}

\item{tripsTable1}{tibble of trips_output (from readTripsTable())}

\item{tripsTable2}{tibble of trips_output (from readTripsTable())}
}
\value{
plots bar chart of changes in modal split
}
\description{
Takes two data frames (from \link{readTripsTable()}), calculates the
changes in mode shares and plots them as a bar chart
Using the parameter unite.modes, specific modes can be renamed into one with the name specified with united.name (by default 'united')
}
