% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{plot_compare_mainmode_sankey}
\alias{plot_compare_mainmode_sankey}
\title{Plot alluvial/sankey diagram of transport mode changes}
\usage{
plot_compare_mainmode_sankey(
  trips_table1,
  trips_table2,
  show.onlyChanges = FALSE,
  unite.columns = character(0),
  united.name = "united"
)
}
\arguments{
\item{show.onlyChanges}{boolean, if it is set to TRUE => sankey diagram only contains changes on axes}

\item{unite.columns}{vector of character string, changes name of all transport modes in the tibble copy to united.name = "united" that matches PATTERNS given in unite.columns}

\item{united.name}{if columns were united, you can specify name for the resulting column in plot}

\item{tripsTable1}{tible of trips_output (from readTripsTable())}

\item{tripsTable2}{tible of trips_output (from readTripsTable())}
}
\value{
Alluvial diagram that represents changes in transport mode distribution of trip tables
}
\description{
Takes two trips_table (from readTripsTable), and collects
changes between transport mode distribution of these tables
to make alluvial diagram from this data
}
\details{
Function calculates number of each transport mode used in
first and second table, and draws plot that represent how
distribution of transport mode has changed (f. e. what part of concrete trasport mode changed to another)
Using parameter unite.columns transport modes that match PATTERN in unite.columns can be united in 1 transport mode type (by default united.name is "united")
Using parameter show.onlyChanges
}
