% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_persons.R
\name{plot_compare_score_boxplot}
\alias{plot_compare_score_boxplot}
\title{Create a boxplot of score differences}
\usage{
plot_compare_score_boxplot(person_tibble1, person_tibble2)
}
\arguments{
\item{person_tibble1}{persons tibble of the base case, can be loaded with read_output_persons.}

\item{person_tibble2}{persons tibble of the policy case, can be loaded with read_output_persons.}
}
\value{
ggplot boxplot of the distribution of the score differences
}
\description{
Takes two tibbles created by \link{read_output_persons}
to calculate the score difference and create a boxplot.
}
