% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{plot_compare_travelwaittime_by_mainmode}
\alias{plot_compare_travelwaittime_by_mainmode}
\title{Bar Chart comparing two runs with main_mode on x-axis and average travel/wait time on y-axis}
\usage{
plot_compare_travelwaittime_by_mainmode(
  trips_table1,
  trips_table2,
  unite.modes = character(0),
  united.name = "united",
  time_format = "minute"
)
}
\arguments{
\item{unite.modes}{vector of character strings,
changes names of chosen modes in the column main_mode to a new chosen name (i.e. drtNorth and drtSouth to drt),
using the function (\link{process_rename_mainmodes})}

\item{united.name}{character string, specifies the name of the united mode}

\item{tripsTable1}{tibble of output_trips (from read_output_trips())}

\item{tripsTable2}{tibble of output_trips (from read_output_trips())}
}
\value{
Bar chart plot comparing average time spent on travel/wait of two runs
}
\description{
Takes two data frames (from \link{read_output_trips()}),
to plot a comparison bar chart of travel and wait times.
Using the parameter unite.modes, specific modes can be renamed into one with the name specified with united.name (by default 'united')
}
