% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{plot_count_scatterplot}
\alias{plot_count_scatterplot}
\title{Creates a Via-Style scatterplot for each run}
\usage{
plot_count_scatterplot(
  joined_frame,
  ll = ~x * 0.8 - 200,
  ul = ~x * 1.2 + 200,
  threshold = 100
)
}
\arguments{
\item{joined_frame}{A tibble from process_join_counts_and_links}

\item{ll}{Formula to calculate lower limit of the quality label 'exact', default = 0.8*x - 200}

\item{ul}{Formula to calculate lower limit of the quality label 'exact', default = 1.2*x + 200}

\item{threshold}{Threshold from which data is scaled to log10.}
}
\value{
An adjustable ggplot scatter plot
}
\description{
Takes a tibble from process_join_counts_and_links to create a scatterplot with counts on the x axis and MATSim DTV on the y axis, colored
by the road type.
Lower and upper limits define the section which is considered an 'exact' estimation. Limits
are defined by custom formulas.
}
