% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{plot_deptime_by_act}
\alias{plot_deptime_by_act}
\title{Plot Departure Time by Activity}
\usage{
plot_deptime_by_act(
  trips_table,
  unite_activities = character(0),
  united_name = "united"
)
}
\arguments{
\item{unite_activities}{vector of character strings, that represent patterns of columns to be united, changes name of all activity types in the tibble copy to united_name = "united" that matches PATTERNS given in unite_columns}

\item{united_name}{character string, if columns were united, you can specify name for the resulting column in plot}

\item{tripsTable}{tibble of trips_output (from \link{read_output_trips()})}
}
\value{
A plotly object representing the line plot of departure time by activity.
}
\description{
The \code{plot_deptime_by_act} function generates a line plot to visualize the departure time of trips by activity in a given trips_table tibble.
It provides insights into the number of trips starting at different times for each activity.
}
