% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{plot_map_trips}
\alias{plot_map_trips}
\title{Plots start and end coordinates of a given output_trips table onto an osm map
XXXX functions uses parameter shape_table?}
\usage{
plot_map_trips(trips_table, crs, optimized = FALSE, shape_table = NULL)
}
\arguments{
\item{trips_table}{tibble of output_trips (from read_output_trips())#'}

\item{crs}{numeric representation of the EPSG code or proj4string for the corresponding coordinate system of the trip coordinates, can be found in network file from output directory of MATSim simulation}

\item{optimized}{bool, by default FALSE and gives interactive plot using leaflet, if TRUE creates image with ggplot}
}
\value{
plot with trips
}
\description{
Plots start and end coordinates of a given output_trips table onto an osm map
XXXX functions uses parameter shape_table?
}
