% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{plot_travelwaittime_mean_barchart}
\alias{plot_travelwaittime_mean_barchart}
\title{Plot travel and wait time for each mode as a bar chart}
\usage{
plot_travelwaittime_mean_barchart(
  trips_table,
  unite.modes = character(0),
  united.name = "united",
  time_format = "minute"
)
}
\arguments{
\item{trips_table}{tibble of output_trips (from \links{read_output_trips()})}

\item{unite.modes}{vector of character strings,
changes names of chosen modes in the column main_mode to a new chosen name (i.e. drtNorth and drtSouth to drt),
using the function (\link{process_rename_mainmodes})}

\item{united.name}{character string, specifies the name of the united mode}
}
\value{
Bar Chart plot of average time spent on travel/wait
}
\description{
Takes the data frame output_trips (from \links{read_output_trips()}),
to plot a bar chart of travel and wait times.
Using the parameter unite.modes, specific modes can be renamed into one with the name specified with united.name (by default 'united')
}
