% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripsOutput.R
\name{plotMapWithTrips}
\alias{plotMapWithTrips}
\title{Plots result of filtered trips on the map (from shape)}
\usage{
plotMapWithTrips(table, crs, optimized = FALSE)
}
\arguments{
\item{table}{tibble of trips_output (from readTripsTable())}

\item{crs}{numeric of EPSG code or proj4string, can be found in network file from output directory of MATSim simulation}

\item{optimized}{bool, by default FALSE and gives interactive plot using leaflet, if TRUE using image with ggplot}
}
\value{
plot with trips
}
\description{
Takes trips_table and
transforms objects to match mutual CRS(network.xml from MATSimOutputDirectory)
result of filtering is plotted on map of shapeTable where green points are startpoints of trip and red points are endpoints of trip
}
