% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripsOutput.R
\name{plotModalShiftBar}
\alias{plotModalShiftBar}
\title{Plot bar chart diagram of transport mode changes}
\usage{
plotModalShiftBar(
  tripsTable1,
  tripsTable2,
  show.onlyChanges = FALSE,
  unite.columns = character(0),
  united.name = "united",
  dump.output.to = matsimDumpOutputDirectory
)
}
\arguments{
\item{unite.columns}{vector of character string, changes name of all transport modes in the tibble copy to united.name = "united" that matches PATTERNS given in unite.columns}

\item{united.name}{if columns were united, you can specify name for the resulting column in plot}

\item{dump.output.to}{folder that saves and configures yaml for simwrapper. folder where png of plot is stored}

\item{tripsTable}{tible of trips_output (from readTripsTable())}
}
\value{
Alluvial diagram that represents changes in transport mode distribution of trip tables
}
\description{
Takes two trips_table (from readTripsTable), and collects
changes between transport mode distribution of these tables
to make bar chart diagram with dodging positioning from this data
}
\details{
Function calculates number of each transport mode used in
first and second table, and draws plot that represent how
distribution of transport mode has changed (f. e. what part of concrete trasport mode changed to another)
Using parameter unite.columns transport modes that match PATTERN in unite.columns can be united in 1 transport mode type (by default united.name is "united")
Using parameter show.onlyChanges
}
