% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripsOutput.R
\name{plotModalSplitPieChart}
\alias{plotModalSplitPieChart}
\title{Plot main_mode as a Pie Chart}
\usage{
plotModalSplitPieChart(
  tripsTable,
  unite.columns = character(0),
  united.name = "united",
  dump.output.to = matsimDumpOutputDirectory
)
}
\arguments{
\item{tripsTable}{tible of trips_output (from readTripsTable())}

\item{unite.columns}{vector of character strings, that represent patterns of columns to be united, changes name of all transport modes in the tibble copy to united.name = "united" that matches PATTERNS given in unite.columns}

\item{united.name}{character string, if columns were united, you can specify name for the resulting column in chart}

\item{dump.output.to}{folder that saves and configures yaml for simwrapper dashboard. folder where png of plot is stored}
}
\value{
Pie Chart plot of transport mode distribution, values given in percents
}
\description{
Takes Table trips_output (from readTripsTable()),
to plot pie chart with with values that represent
percentage of using transport modes from trips
}
\details{
Function automatically detects transport_modes from table
and plots pie chart.
Using parameters unite.columns, specific columns could be given, to unite them in 1 mode with the name united.name(by default 'united')
}
