% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{plotTripDistancedByType}
\alias{plotTripDistancedByType}
\title{Bar Chart with tripType on x-axis and travelled distance on y-axis}
\usage{
plotTripDistancedByType(
  tripsTable,
  shapeTable,
  crs,
  dump.output.to = matsimDumpOutputDirectory
)
}
\arguments{
\item{tripsTable}{tible of trips_output (from readTripsTable())}

\item{shapeTable}{sf object(data.frame with geometries), can be received by using st_read(path_to_geographical_file)}

\item{crs}{numeric of EPSG code or proj4string, can be found in network file from output directory of MATSim simulation}

\item{dump.output.to}{folder that saves and configures yaml for simwrapper dashboard. folder where png of plot is stored}
}
\value{
Bar Chart plot of distance traveled by type
}
\description{
Takes Table trips_output (from readTripsTable()),
to plot bar chart with with values that represent
travelled distance of each tripType related to the shapeTable
}
