% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripsOutput.R
\name{plotTripTypesPieChart}
\alias{plotTripTypesPieChart}
\title{Plots distribution of every type of trips(inside, outside, origin and destinating) in Pie Chart}
\usage{
plotTripTypesPieChart(table, shapeTable, crs)
}
\arguments{
\item{table}{tibble of trips_output (from readTripsTable())}

\item{shapeTable}{sf object(data.frame with geometries), can be received by using st_read(path_to_geographical_file)}

\item{crs}{numeric of EPSG code or proj4string, can be found in network file from output directory of MATSim simulation}
}
\value{
plot with percentage of each type of trips
}
\description{
Plots distribution of every type of trips(inside, outside, origin and destinating) in Pie Chart
}
