% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simwrapper.R
\name{prepareComparisonSimwrapperDashboardFromTable}
\alias{prepareComparisonSimwrapperDashboardFromTable}
\title{Creates comparison dashboard for the 2 given tables or folder with data}
\usage{
prepareComparisonSimwrapperDashboardFromTable(
  table1,
  table2,
  dump.output.to = matsimDumpOutputDirectory,
  append = FALSE
)
}
\arguments{
\item{table1}{trips_output tibble from readTripsTable()}

\item{table2}{trips_output tibble from readTripsTable()}

\item{dump.output.to}{folder that saves and configures yaml for simwrapper dashboard and all plots using functions:
plotModalSplitBarChart(),plotModalSplitPieChart(),plotModalShift().}

\item{append}{specifies if the ouput folder should be erased before creating}
}
\value{
generates folder with content for simwrapper out of trips table
}
\description{
Creates comparison dashboard for the 2 given tables or folder with data
}
