% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripsOutput.R
\name{prepareSimwrapperDashboardFromTable}
\alias{prepareSimwrapperDashboardFromTable}
\title{Creates dashboard for the given table or folder with data}
\usage{
prepareSimwrapperDashboardFromTable(
  table,
  dump.output.to = matsimDumpOutputDirectory,
  append = FALSE
)
}
\arguments{
\item{table}{trips_output tibble from readTripsTable()}

\item{append}{specifies if the ouput folder should be erased before creating}
}
\value{
plot that contains every trip with defined trip type
}
\description{
Creates dashboard for the given table or folder with data
}
