% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{process_append_distcat}
\alias{process_append_distcat}
\title{Adds additional \strong{dist_cat} column representing category of distance traveled based on \strong{distances_array} parameter.}
\usage{
process_append_distcat(
  trips_table,
  distances_array = c(1000, 2000, 5000, 10000, 20000, 50000, 1e+05)
)
}
\arguments{
\item{trips_table}{tibble of trips_output (from \link{read_output_trips})}

\item{distances_array}{numeric vector, represents segments for distance categories ordered (in meters)}
}
\description{
Categorize column of distances based on \strong{distances_array}, each dist_cat receives relational segment from array.\cr
For example trip distance traveled is 1500, and distances array is (1000,2000), then this category is "1000-2000".
}
