% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{process_get_crs_from_config}
\alias{process_get_crs_from_config}
\title{Reads the coordinate reference system from a MATSim output directory (output_config.xml)}
\usage{
process_get_crs_from_config(config_path)
}
\arguments{
\item{config_path}{specifies path to configuration file}
}
\value{
EPSG code of coordinate reference system
}
\description{
Reads the coordinate reference system from a MATSim output directory (output_config.xml)
}
