% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{process_get_mainmode_distribution}
\alias{process_get_mainmode_distribution}
\title{Calculates the count or percentage for each mode}
\usage{
process_get_mainmode_distribution(trips_table, percentage = FALSE)
}
\arguments{
\item{trips_table}{tibble of output_trips (from read_output_trips())}

\item{percentage}{boolean, standard value FALSE (calculates the count)}
}
\value{
tibble including the count or percentage
}
\description{
Calculates the count or percentage for each mode in the column main_mode
}
