% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{process_get_od_matrix}
\alias{process_get_od_matrix}
\title{Creates an origin/destination matrix either in conventional form (row names = origin, column names = destination)
or for simwrapper (origin and destination as columns)}
\usage{
process_get_od_matrix(
  trips_table,
  shape_path,
  crs,
  dump.output.to = matsimDumpOutputDirectory,
  simwrapper = FALSE,
  colnames = "numeric",
  outer = FALSE
)
}
\arguments{
\item{crs}{numeric, coordinate system in the form of the EPSG code or proj4string, can be found in the MATSim network file}

\item{dump.output.to}{string, path to a folder to save the .csv file}

\item{simwrapper}{boolean, creates output in the format used for simwrapper if the path for the shapefile is specified}

\item{colnames}{string, column names can be specified (i.e. to fit the shape file), if not they are numbered}

\item{outer}{boolean, determines if flows outside of the shapefile are used, standard value is FALSE}

\item{tripsTable}{tibble of trips_output (from \link{readTripsTable})}

\item{shapePath}{string, full path to the shapefile (.shp) (shape files are made up of several files with the same name and the folder also needs to include a .dbf file)}
}
\value{
tibble of origin/destination matrix
}
\description{
Creates an origin/destination matrix either in conventional form (row names = origin, column names = destination)
or for simwrapper (origin and destination as columns)
}
