% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{process_get_travdistance_distribution}
\alias{process_get_travdistance_distribution}
\title{Calculates the travel distance distribution by main mode}
\usage{
process_get_travdistance_distribution(trips_table, euclidean = FALSE)
}
\arguments{
\item{trips_table}{tibble of output_trips (from \code{\link{read_output_trips}})}

\item{euclidean}{boolean, standard value FALSE, if set to TRUE uses euclidean distance.}
}
\value{
tibble with two columns, main_mode containing the mode and avg_dist containing the average distance traveled by that mode
}
\description{
The function process_get_travdistance_distribution calculates the distribution of travel distances based on main modes in a given trips_table tibble.
It provides insights into the average distance traveled for each main mode.
The distance can also be calculated by using the euclidean distance between origin and destination. To use the euclidean distance set the parameter euclidean as TRUE.
}
