% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{process_join_counts_and_links}
\alias{process_join_counts_and_links}
\title{Join counts and linkstats to the network, creating a tibble into memory}
\usage{
process_join_counts_and_links(
  counts,
  network,
  link_stats,
  network_modes = c("car"),
  aggr_to = c("day", "hour"),
  earliest = 0,
  latest = 86400
)
}
\arguments{
\item{counts}{tibble of counts from \link{read_counts}}

\item{network}{list with network nodes and links}

\item{link_stats}{List with linkstats tibbles}

\item{network_modes}{Vector with network modes that will be analyzed, default is "car".}

\item{aggr_to}{Determines if data should be aggregated into hourly bins or as daily traffic volume, can either be "day" or "hour"}

\item{earliest}{Integer. Lower limit to filter link stats by time, default = 0.}

\item{latest}{Integer. Upper limit to filter link stats by time, default = 86400 (midnight).}
}
\value{
Long-format tibble with MATSim link id as key ("loc_id"), traffic volumes from MATSim runs and link type
}
\description{
Function to join counts, linkstats and network links. Data can be aggregated
and filtered by time or mode.
}
