% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{process_rename_category}
\alias{process_rename_category}
\title{XXXX}
\usage{
process_rename_category(
  trips_table,
  unite_template = character(0),
  united_name = "united",
  column = "main_mode"
)
}
\description{
XXXX
}
