% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{process_rename_mainmodes}
\alias{process_rename_mainmodes}
\title{Replace Matching Categories of main_mode with United Name in Tibble}
\usage{
process_rename_mainmodes(
  trips_table,
  unite_modes = character(0),
  united_name = "united"
)
}
\arguments{
\item{trips_table}{tible of output_trips (from \link{\code{read_output_trips}})}

\item{unite_modes}{vector of character string, representing pattern to match categories}

\item{united_name}{character string, to represent new category for all occurrences matched unite_modes}
}
\value{
tibble with new main_mode representation
}
\description{
The process_rename_mainmodes function takes a tibble, trips_table,
as input and replaces all occurrences in column main_mode of categories
that match a given template with another name provided in united_name.
The function provides the flexibility to specify multiple category templates
to be united into a single name.
}
