% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{processDtvEstimationQuality}
\alias{processDtvEstimationQuality}
\title{Categorize DTV deviation and aggregate data}
\usage{
processDtvEstimationQuality(joinedFrame)
}
\arguments{
\item{joinedFrame}{A tibble from mergeCountsAndLinks}
}
\value{
A long-format tibble, which contains share of estimation quality for each scenario and link type
}
\description{
Takes a tibble from mergeCountsAndLinks.
Deviation between count volumes and Linkstats is calculated
(e.g. deviation of 1.2 means 20 percent more DTV in MATSim than in counts) and
categorized.
Data will be aggregated for each run and link type.
Can be used to visualize model quality by link type and to compare several runs.
}
