% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{processLinkStatsForScatterPlot}
\alias{processLinkStatsForScatterPlot}
\title{Prepares Linkstats and counts for VIA-style scatter plot}
\usage{
processLinkStatsForScatterPlot(joinedFrame)
}
\arguments{
\item{joinedFrame}{a tibble, created from function mergeCountsAndLinks}
}
\value{
a long-format tibble, ready to for plotting
}
\description{
Takes a tibble from mergeCountsAndLinks and prepare data for a VIA-style
scatterplot for one or more runs.
Only DTV values for mode 'car' is processed.
}
\details{
Pattern "highway." is removed from link type, also link type is factored and
convert tibble to long-format
}
