% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{read_config}
\alias{read_config}
\title{Load MATSIM config file into Memory}
\usage{
read_config(input_path = ".", n_max = Inf)
}
\arguments{
\item{input_path}{character string, path to matsim output directory or http link to the file.}

\item{n_max}{integer, maximum number of lines to read within output_trips}
}
\value{
tibble of trips_output
}
\description{
Loads a MATSim xml config from file or archive,
creating a list with parameters as in xml file
}
