% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{read_linkstats}
\alias{read_linkstats}
\title{Load a MATSim linkstats file into memory}
\usage{
read_linkstats(run_ID, input_path, sample_size = 0.25)
}
\arguments{
\item{run_ID}{character string, ID to tag columns with DTV}

\item{input_path}{character string, file path to the .csv or .tsv file. Must be comma separated.}

\item{sample_size}{optional, double, sample size of the MATSim scenario to scale DTV values, standard value is 0.25}
}
\value{
Tibble with link stats for each qsim mode
}
\description{
Loads a linkstats tsv file created from the LinkStats class
as a dataframe into memory.
Counts can be provided in any time bin.
Counts can be provided for any qsim mode. The argument network_modes is used to
select and filter the columns.
}
