% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{read_network}
\alias{read_network}
\title{Load MATSim network into memory}
\usage{
read_network(filename)
}
\arguments{
\item{filename}{File to load. Can be XML or gzipped XML}
}
\value{
"nodes" and "links" tibbles in a list object.
}
\description{
Loads a MATSim XML network file, creating a nodes tibble and a links tibble.
Any node and link attribute records in the network are stored as
additional columns in the respective node and link tibbles.
}
\details{
The links table is automatically joined with the nodes table so that
node x/y coordinates (and any other node attributes) are available in the
links table without additional processing.
}
