% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_persons.R
\name{read_output_persons}
\alias{read_output_persons}
\title{Load MATSim output_persons table into memory}
\usage{
read_output_persons(input_path = ".", n_max = Inf)
}
\arguments{
\item{input_path}{is a character string, path to the local MATSim output directory, to the persons csv directly, or a http link to the file.}

\item{n_max}{integer, maximum number of lines to read within output_persons}
}
\value{
tibble of output_persons
}
\description{
Loads a MATSim CSV output_persons from file or archive,
creating a tibble with columns as in csv file
copied + adopted code from readTripsTable in tripsOutput.R
}
