% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{read_output_trips}
\alias{read_output_trips}
\title{Load MATSIM output_trips table into memory}
\usage{
read_output_trips(input_path = ".", n_max = Inf)
}
\arguments{
\item{input_path}{character string, path to the MATSim output directory or http link to the file.}

\item{n_max}{optional, integer, maximum number of lines to read, standard value is Inf}
}
\value{
tibble of trips_output
}
\description{
Loads a MATSim CSV output_trips from file or archive,
creating a tibble with columns as in csv file
}
