% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personsOutput.R
\name{readPersonsTable}
\alias{readPersonsTable}
\title{Load MATSim output_persons table into memory}
\usage{
readPersonsTable(input_path = ".")
}
\arguments{
\item{input_path}{is a character string, path to the local MATSim output directory, to the persons csv directly, or a http link to the file.}
}
\value{
tibble of output_persons
}
\description{
Loads a MATSim CSV output_persons from file or archive,
creating a tibble with columns as in csv file
copied + adopted code from readTripsTable in tripsOutput.R
}
