% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripsOutput.R
\name{readTripsTable}
\alias{readTripsTable}
\title{Load MATSIM output_trips table into Memory}
\usage{
readTripsTable(pathToMATSimOutputDirectory = ".")
}
\arguments{
\item{pathTOMATSimOutputDirectory}{matsim output directory or link to the file.}
}
\value{
tibble of trips_output
}
\description{
Loads a MATSim CSV output_trips file or archive,
creating a tibble with columns as in csv file
}
