% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{symlog_trans}
\alias{symlog_trans}
\title{A function to create symlog scaling for a plot}
\usage{
symlog_trans(base = 10, thr = 1, scale = 1)
}
\arguments{
\item{base}{base for log}

\item{thr}{threshold from which data is scaled to log}

\item{scale}{scale}
}
\description{
Can be used to symlog scale the axis of a ggplot object. Is called in plot_count_scatterplot.
Note that this function is taken from Stackoverflow!
For more informations, see the thread here:
https://stackoverflow.com/questions/14613355/how-to-get-something-like-matplotlibs-symlog-scale-in-ggplot-or-lattice
}
