% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personsOutput.R
\name{transform_persons_sf}
\alias{transform_persons_sf}
\title{Transforms persons_table tibble (from readPersonsTable) from tibble to sf (table with attribute features and geometry feature)}
\usage{
transform_persons_sf(
  persons,
  crs = "EPSG:25832",
  filter_shp = NULL,
  first_act_type_filter = NULL
)
}
\arguments{
\item{persons}{table containing persons (from readPersonsTable)}

\item{crs}{numeric of EPSG code or proj4string, can be found in network file from output directory of MATSim simulation}

\item{filter_shp}{(optional) if provided, will filter persons_table to only include first_activity_locations within given shape file}

\item{first_act_type_filter}{(optional) if provided, will filter persons_table to only include agents whose day starts with a certain activity type (e.g. home)}
}
\value{
sf object (data.frame with geometries depending to geometry.type)
}
\description{
Takes first activity location of person turns it into simple feature POINT geometry
}
