% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{compareBasePolicyOutput}
\alias{compareBasePolicyOutput}
\title{Chooses a function to compare output_trips from the folders.
baseFolder contains all base outputs, policyFolder contains all policy outputs.}
\usage{
compareBasePolicyOutput(
  baseFolder,
  policyFolder,
  dump.output.to = matsimDumpOutputDirectory
)
}
\arguments{
\item{baseFolder}{specifies data source folder with multiple base output_trips}

\item{policyFolder}{specifies data source folder with multiple policy output_trips}

\item{dump.output.to}{that saves result of all comparisons between each base and each policy.
For now it creates plotModalShiftBar() for the output_trips}
}
\value{
list of tibbles, list of all base and policy output_trips as tibble
}
\description{
Chooses a function to compare output_trips from the folders.
baseFolder contains all base outputs, policyFolder contains all policy outputs.
}
