% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripsOutput.R
\name{deriveODMatrix}
\alias{deriveODMatrix}
\title{Creates an instance of ODMatrix(origin/destination) in conventional form or for the simwrapper}
\usage{
deriveODMatrix(
  tripsTable,
  shape,
  crs,
  dump.output.to = matsimDumpOutputDirectory,
  simwrapper = FALSE
)
}
\arguments{
\item{tripsTable}{table of output trips(from readTripsTable)}

\item{crs}{numeric of EPSG code or proj4string, can be found in network file from output directory of MATSim simulation}

\item{dump.output.to}{that saves result of all comparisons between each base and each policy.
For now it creates plotModalShiftBar() for the output_trips}

\item{shapeTable}{sf object(data.frame with geometries), can be received by using st_read(path_to_geographical_file)}
}
\value{
list of tibbles, list of all base and policy output_trips as tibble
}
\description{
Creates an instance of ODMatrix(origin/destination) in conventional form or for the simwrapper
}
