% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripsOutput.R
\name{deriveODMatrix}
\alias{deriveODMatrix}
\title{Creates an instance of ODMatrix(origin/destination) in conventional form or for the simwrapper}
\usage{
deriveODMatrix(
  tripsTable,
  shapePath,
  crs,
  dump.output.to = matsimDumpOutputDirectory,
  simwrapper = FALSE,
  colnames = "numeric",
  outer = FALSE
)
}
\arguments{
\item{tripsTable}{table of output trips(from readTripsTable) or path to trips_output file}

\item{shapePath}{full path to shapefile (if simwrapper TRUE, folder with shapeFile should contain also .dbf with the same name)}

\item{crs}{numeric of EPSG code or proj4string, can be found in network file from output directory of MATSim simulation}

\item{dump.output.to}{path to a folder to save csv file of ODMatrix}

\item{simwrapper}{create output in a simwrapper form if set to path of the shapefile}

\item{colnames}{if the specific shapefile contains known columns, they could be specified as name for columns OD. If not given then they get numeric values}

\item{outer}{logical that represent if the table should contain outside flow of the shape, it isn't}
}
\value{
tibble of origin/destination matrix
}
\description{
Creates an instance of ODMatrix(origin/destination) in conventional form or for the simwrapper
}
