% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripsOutput.R
\name{generateXYHexagonYaml}
\alias{generateXYHexagonYaml}
\title{Creates/adds XY hexagon definition of output trips to a summary dashboard}
\usage{
generateXYHexagonYaml(
  pathToOutputTrips,
  crs,
  dump.output.to = matsimDumpOutputDirectory
)
}
\arguments{
\item{pathToOutputTrips}{specifies path to a data source of output_trips}

\item{crs}{specifies an coordinate reference system of output_trips}

\item{dump.output.to}{path to an folder with dashboard}
}
\value{
changed file
}
\description{
Creates/adds XY hexagon definition of output trips to a summary dashboard
}
