% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\docType{package}
\name{matsimr-deprecated}
\alias{matsimr-deprecated}
\alias{plotModalSplitPieChart}
\alias{plotModalSplitBarChart}
\title{Plot main_mode distribution as a Pie Chart}
\usage{
plotModalSplitPieChart(
  tripsTable,
  unite.columns = character(0),
  united.name = "united",
  dump.output.to = matsimDumpOutputDirectory,
  only.files = FALSE
)
}
\description{
These functions are provided for compatibility with older version of
the yourPackageName package.  They may eventually be completely
removed.
}
\section{Details}{

\tabular{rl}{
  \code{plot_mainmode_piechart} \tab now a synonym for \code{\link{plotModalSplitPieChart}}\cr
  \code{plot_mainmode_barchart} \tab now a synonym for \code{\link{plotModalSplitBarChart}}\cr
}
}

