% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\docType{package}
\name{matsimr-deprecated}
\alias{matsimr-deprecated}
\alias{plotModalSplitPieChart}
\alias{plotModalSplitBarChart}
\title{Deprecated function(s) in the matsimr package}
\usage{
plotModalSplitPieChart(
  tripsTable,
  unite.columns = character(0),
  united.name = "united",
  dump.output.to = matsimDumpOutputDirectory,
  only.files = FALSE
)

plotModalSplitBarChart(
  tripsTable,
  unite.columns = character(0),
  united.name = "united",
  dump.output.to = matsimDumpOutputDirectory,
  only.files = FALSE
)
}
\description{
These functions are provided for compatibility with older version of
the matsimr package.  They may eventually be completely
removed.

These functions are provided for compatibility with older version of
the matsimr package.  They may eventually be completely
removed.
}
\details{
Test of usage
}
\section{Details}{

\tabular{rl}{
  \code{plotModalSplitPieChart} \tab now a synonym for \code{\link{plot_mainmode_piechart}}\cr
  \code{plotModalSplitBarChart} \tab now a synonym for \code{\link{plot_mainmode_barchart}}\cr
}


\tabular{rl}{
  \code{plotModalSplitPieChart} \tab now a synonym for \code{\link{plot_mainmode_piechart}}\cr
  \code{plotModalSplitBarChart} \tab now a synonym for \code{\link{plot_mainmode_barchart}}\cr
}
}

