% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\docType{package}
\name{matsimr-deprecated}
\alias{matsimr-deprecated}
\alias{plotModalSplitPieChart}
\alias{plotModalSplitBarChart}
\alias{readTripsTable}
\alias{plotAverageTravelWait}
\alias{compareAverageTravelWait}
\alias{plotModalDistanceDistribution}
\alias{appendDistanceCategory}
\title{Deprecated function(s) in the matsimr package}
\usage{
plotModalSplitPieChart(
  tripsTable,
  unite.columns = character(0),
  united.name = "united",
  dump.output.to = matsimDumpOutputDirectory,
  only.files = FALSE
)

plotModalSplitBarChart(
  tripsTable,
  unite.columns = character(0),
  united.name = "united",
  dump.output.to = matsimDumpOutputDirectory,
  only.files = FALSE
)

readTripsTable(input_path = ".", n_max = Inf)

plotAverageTravelWait(
  tripsTable,
  unite.columns = character(0),
  united.name = "united",
  dump.output.to = matsimDumpOutputDirectory,
  only.files = FALSE
)

compareAverageTravelWait(
  tripsTable1,
  tripsTable2,
  unite.columns = character(0),
  united.name = "united",
  dump.output.to = matsimDumpOutputDirectory,
  only.files = FALSE
)

plotModalDistanceDistribution(
  tripsTable,
  unite.columns = character(0),
  united.name = "united",
  dump.output.to = matsimDumpOutputDirectory,
  only.files = FALSE
)

appendDistanceCategory(tripsTable)
}
\arguments{
\item{tripsTable}{tible of trips_output (from readTripsTable())}

\item{unite.columns}{vector of character strings, that represent patterns of columns to be united, changes name of all transport modes in the tibble copy to united.name = "united" that matches PATTERNS given in unite.columns}

\item{united.name}{character string, if columns were united, you can specify name for the resulting column in chart}

\item{dump.output.to}{folder that saves and configures yaml for simwrapper dashboard. folder where png of plot is stored}

\item{only.files}{boolean, that represent if plotting inside project is needed, by default FALSE - means function gives out a plot by plot_ly}

\item{input_path}{character string, path to matsim output directory or http link to the file.}

\item{n_max}{integer, maximum number of lines to read within output_trips}
}
\value{
tibble of trips_output

Bar Chart plot of count of trips among distance travelled
}
\description{
These functions are provided for compatibility with older version of
the matsimr package.  They may eventually be completely
removed.

Loads a MATSim CSV output_trips from file or archive,
creating a tibble with columns as in csv file

Takes Table trips_output (from readTripsTable()),
to plot bar chart with with values that represent
number of trips ~ distance travelled
Using parameters unite.columns, specific columns could be given, to unite them in 1 mode with the name united.name(by default 'united')
}
\section{Details}{

\tabular{rl}{
  \code{plotModalSplitPieChart} \tab now a synonym for \code{\link{plot_mainmode_piechart}}\cr
  \code{plotModalSplitBarChart} \tab now a synonym for \code{\link{plot_mainmode_barchart}}\cr
  \code{readTripsTable} \tab now a synonym for \code{\link{read_output_trips}}\cr
}
}

