% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{mergeCountsAndLinks}
\alias{mergeCountsAndLinks}
\title{Load Counts, a limited number of Linkstats and Network links as joined tibble into memory}
\usage{
mergeCountsAndLinks(
  counts,
  network,
  linkStats,
  networkModes = c("car"),
  aggr = TRUE,
  earliest = 0,
  latest = 86400
)
}
\arguments{
\item{counts}{Tibble with counts data}

\item{network}{Tibble with network nodes and links}

\item{linkStats}{List with link stats tibbles}

\item{networkModes}{a vector with network modes, which are needed for analysis}

\item{aggr}{Boolean, if data should be aggregated}

\item{earliest}{Lower limit to filter link stats by time, default is 0}

\item{latest}{Upper limit to filter link stats by time, default is 86400 (midnight)}
}
\value{
Long-format tibble with MATSim link id as key ("loc_id"), traffic volume from MATSim runs and link type
}
\description{
Function to join counts, network links and several matsim link stats. Data can be aggregated
and filtered by time or network mode.
}
