% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{mergeCountsAndLinks}
\alias{mergeCountsAndLinks}
\title{Load Counts, a limited number of Linkstats and Network links as joined tibble into memory}
\usage{
mergeCountsAndLinks(
  countsFilePath,
  networkFilePath,
  linkStatsList,
  sampleSize = NA,
  outputFilePath = NA
)
}
\arguments{
\item{countsFilePath}{Path to MATSim counts file}

\item{networkFilePath}{Path to MATSim network file}

\item{linkStatsList}{List with filepaths to Linkstats, uses the runId as key, and filepath as value}

\item{sampleSize}{sample size of the scenario between 0 and 1 to rescale DTV values}

\item{outputFilePath}{output file path to write joined tibble to .csv file}
}
\value{
Tibble with MATSim link id as key ("loc_id"), DTV from MATSim runs and link type
}
\description{
Function to import and join MATSim Counts, Linkstats and network link types
Linkstats will be aggregated to DTV. An additional sampleSize parameter can be
used to scale DTV values.
Joined tibble can be written to a .csv file to provide processed data for further uses
and to save some run time.
}
\examples{

runList = list("run_1" = "path/to/linkstats_1",
              "run_2" = "path/to/linkstats_2"
              )

mergeCountsAndLinks(
 countsFilePath = "path/to/counts.xml",
 networkFilePath = "path/to/matsim-network.xml.gz",
 linkStatsList = runList,
 sampleSize = 0.25,
 outputFilePath = "path/to/desired/output/file.csv"
 )


}
