% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{plot_arrtime_by_act}
\alias{plot_arrtime_by_act}
\title{Plot Arrival Time by Activity}
\usage{
plot_arrtime_by_act(
  trips_table,
  unite_activities = character(0),
  united_name = "united"
)
}
\arguments{
\item{trips_table}{tibble of trips_output (from \link{read_output_trips()})}

\item{unite_activities}{vector of character strings, that represent patterns of columns to be united, changes name of all activity types in the tibble copy to united_name = "united" that matches PATTERNS given in unite.columns}

\item{united_name}{character string, if columns were united, you can specify name for the resulting column in plot}
}
\value{
A plotly object representing the line plot of arrival time by activity.
}
\description{
\strong{plot_arrtime_by_act} - Takes Table trips_output (from \link{read_output_trips}()),
to make line plot with with values that represent
count of destination activities for a specific arrival time
Using parameters unite.columns, specific columns could be given, to unite them in 1 mode with the name united_name(by default 'united')
}
