% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{plot_arrtime_by_act}
\alias{plot_arrtime_by_act}
\title{Deprecated function(s) in the matsimr package}
\usage{
plot_arrtime_by_act(
  trips_table,
  unite_activities = character(0),
  united_name = "united"
)
}
\arguments{
\item{trips_table}{tibble of trips_output (from readTripsTable())}

\item{unite.columns}{vector of character strings, that represent patterns of columns to be united, changes name of all activity types in the tibble copy to united.name = "united" that matches PATTERNS given in unite.columns}

\item{united.name}{character string, if columns were united, you can specify name for the resulting column in plot}

\item{dump.output.to}{folder that saves and configures yaml for simwrapper dashboard. folder where png of plot is stored}

\item{only.files}{boolean, that represent if plotting inside project is needed, by default FALSE - means function gives out a plot by plot_ly}
}
\value{
\strong{plot_arrtime_by_act} - Line plot with arrival time x-axis and number end activities on y-axis
}
\description{
\strong{plot_arrtime_by_act} - Takes Table trips_output (from readTripsTable()),
to make line plot with with values that represent
count of destination activities for a specific arrival time
Using parameters unite.columns, specific columns could be given, to unite them in 1 mode with the name united.name(by default 'united')
}
