% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{plot_compare_count_by_spatialcat_barchart}
\alias{plot_compare_count_by_spatialcat_barchart}
\title{Comparison bar chart with trip type on x-axis and count on y-axis}
\usage{
plot_compare_count_by_spatialcat_barchart(
  trips_table1,
  trips_table2,
  shape_table,
  crs,
  dump.output.to = matsimDumpOutputDirectory
)
}
\arguments{
\item{crs}{numeric representation of the EPSG code or proj4string for the corresponding coordinate system
of the trip coordinates, can be found in network file from output directory of MATSim simulation}

\item{tripsTable1}{tibble of output_trips (from read_output_trips())}

\item{tripsTable2}{tibble of output_trips (from read_output_trips())}

\item{shapeTable}{sf object(data.frame with geometries), can be created using st_read()}
}
\value{
Bar chart comparing number of each trip type for two runs
}
\description{
Takes two data frames (from \link{read_output_trips()}) to plot a bar chart comparing the number of each trip type (originating, destinating, inside, outside).
The shape file is used to categorize the trips.
}
