% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{plot_compare_count_by_spatialcat_barchart}
\alias{plot_compare_count_by_spatialcat_barchart}
\title{Plot Comparison of Count by Spatial Category (Bar Chart)}
\usage{
plot_compare_count_by_spatialcat_barchart(
  trips_table1,
  trips_table2,
  shape_table,
  crs
)
}
\arguments{
\item{trips_table1}{tible of trips_output (from \link{read_output_trips})}

\item{trips_table2}{tible of trips_output (from \link{read_output_trips})}

\item{shape_table}{A spatial shapefile or spatial polygons data frame used to create the spatial categories.}

\item{crs}{numeric representation of the EPSG code or proj4string for the corresponding coordinate system of the trip coordinates,
can be found in network file from output directory of MATSim simulation}
}
\value{
A plotly object representing the bar chart of the comparison of count by spatial category.
}
\description{
The \code{plot_compare_count_by_spatialcat_barchart}
function generates a bar chart to compare the count of trips between two trips_table
tibbles across different spatial categories.
It provides insights into the distribution of trips across spatial categories for each table.
}
