% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{plot_compare_distcat_by_mainmode_barchart}
\alias{plot_compare_distcat_by_mainmode_barchart}
\title{Bar Chart with distance travelled on x-axis and difference of number of trips on y-axis}
\usage{
plot_compare_distcat_by_mainmode_barchart(
  trips_table1,
  trips_table2,
  unite.columns = character(0),
  united.name = "united",
  dist_column = "dist_cat",
  distances_array = c(1000, 2000, 5000, 10000, 20000, 50000, 1e+05)
)
}
\arguments{
\item{trips_table1}{tible of trips_output (from readTripsTable()), number of trips of this table will be extracted from number of trips of tripsTable1}

\item{trips_table2}{tible of trips_output (from readTripsTable()), from number of trips of this table number of trips of tripsTable1 will be extracted}

\item{unite.columns}{vector of character strings, that represent patterns of columns to be united, changes name of all transport modes in the tibble copy to united.name = "united" that matches PATTERNS given in unite.columns}

\item{united.name}{character string, if columns were united, you can specify name for the resulting column in chart}

\item{dump.output.to}{folder that saves and configures yaml for simwrapper dashboard. folder where png of plot is stored}

\item{only.files}{boolean, that represent if plotting inside project is needed, by default FALSE - means function gives out a plot by plot_ly}
}
\value{
Bar Chart plot of count of trips among distance travelled
}
\description{
Takes 2 Tables trips_output (from readTripsTable()),
to plot bar chart with with values that represent
difference of number of trips between tripsTable2 and tripsTable1 ~ distance travelled
Using parameters unite.columns, specific columns could be given, to unite them in 1 mode with the name united.name(by default 'united')
}
