% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{plot_compare_mainmode_sankey}
\alias{plot_compare_mainmode_sankey}
\title{Plot alluvial/sankey diagram of transport mode changes}
\usage{
plot_compare_mainmode_sankey(
  trips_table1,
  trips_table2,
  show_onlychanges = FALSE,
  unite_modes = character(0),
  united_name = "united"
)
}
\arguments{
\item{trips_table1}{tibble of trips_output (from read_output_trips())}

\item{trips_table2}{tibble of trips_output (from read_output_trips())}

\item{show_onlychanges}{boolean, if it is set to TRUE the sankey diagram only shows the mode shift}

\item{unite_modes}{optional, vector of character strings,
changes names of chosen modes in the column main_mode to a new chosen name (i.e. drtNorth and drtSouth to drt),
using the function (\link{process_rename_mainmodes})}

\item{united_name}{optional, character string, specifies the name of the united mode}
}
\value{
Alluvial diagram that represents changes in transport mode distribution
}
\description{
Takes two tibbles from \link{read_output_trips} and compares the mode choice for each agent and summarizes the results, showing the modal shift.
Using the parameter unite_modes, specific modes can be renamed into one with the name specified with united_name (by default 'united')
The parameter show.onlyChanges allows the visualization of only the mode shift (excluding the trips that do not change mode). Standard value is FALSE.
}
