% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{plot_compare_travelwaittime_by_mainmode_barchart}
\alias{plot_compare_travelwaittime_by_mainmode_barchart}
\title{Create comparison bar chart with main_mode on x-axis and average travel/wait time on y-axis}
\usage{
plot_compare_travelwaittime_by_mainmode_barchart(
  trips_table1,
  trips_table2,
  unite_modes = character(0),
  united_name = "united",
  time_format = "minute"
)
}
\arguments{
\item{trips_table1}{tibble of trips_output (from \link{read_output_trips})}

\item{trips_table2}{tibble of trips_output (from \link{read_output_trips})}

\item{unite_modes}{optional, vector of character strings,
changes names of chosen modes in the column main_mode to a new chosen name (i.e. drtNorth and drtSouth to drt),
using the function (\link{process_rename_mainmodes})}

\item{united_name}{optional, character string, specifies the name of the united mode}

\item{time_format}{character string, indicating the desired time format for the output.
The options are "minute", "hour", or "second". The default is "minute".}
}
\value{
Bar chart plot of average travel/wait time
}
\description{
Takes two tibbles from \link{read_output_trips},
to plot a comparison bar chart of the traveling/waiting time.
Using the parameter unite_modes, specific modes can be renamed into one with the name specified with united_name (by default 'united')
}
