% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{plot_distance_by_mainmode_barchart}
\alias{plot_distance_by_mainmode_barchart}
\title{Bar chart with average distance traveled for each mode on x-axis and number of trips on y-axis}
\usage{
plot_distance_by_mainmode_barchart(
  trips_table,
  unite.modes = character(0),
  united.name = "united",
  euclidean = FALSE
)
}
\arguments{
\item{trips_table}{tibble of trips_output (from readTripsTable())}

\item{unite.modes}{vector of character strings,
changes names of chosen modes in the column main_mode to a new chosen name (i.e. drtNorth and drtSouth to drt),
using the function (\link{process_rename_mainmodes})}

\item{united.name}{character string, specifies the name of the united mode}
}
\value{
Bar Chart plot of distance traveled per mode
}
\description{
Takes the data frame trips_output (from \link{readTripsTable()}),
to plot a bar chart of the average distance traveled for each main mode,
Using the parameter unite.modes, specific modes can be renamed into one with the name specified with united.name (by default 'united')
}
