% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{plot_distance_by_mainmode_barchart}
\alias{plot_distance_by_mainmode_barchart}
\title{Bar Chart with distance travelled on x-axis and number of trips on y-axis}
\usage{
plot_distance_by_mainmode_barchart(
  trips_table,
  unite.columns = character(0),
  united.name = "united",
  euclidean = FALSE
)
}
\arguments{
\item{trips_table}{tible of trips_output (from readTripsTable())}

\item{unite.columns}{vector of character strings, that represent patterns of columns to be united, changes name of all transport modes in the tibble copy to united.name = "united" that matches PATTERNS given in unite.columns}

\item{united.name}{character string, if columns were united, you can specify name for the resulting column in chart#'}
}
\value{
Bar Chart plot of distance traveled per mode
}
\description{
Takes Table trips_output (from readTripsTable()),
to plot bar chart with with values that represent
average distance traveled ~ main mode used
Using parameters unite.columns, specific columns could be given, to unite them in 1 mode with the name united.name(by default 'united')
}
