% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{plot_distance_by_spatialcat_barchart}
\alias{plot_distance_by_spatialcat_barchart}
\title{Bar chart with trip type on x-axis and traveled distance on y-axis}
\usage{
plot_distance_by_spatialcat_barchart(
  trips_table,
  shape_table,
  crs,
  euclidian = FALSE
)
}
\arguments{
\item{crs}{numeric representation of the EPSG code or proj4string for the corresponding coordinate system
of the trip coordinates, can be found in network file from output directory of MATSim simulation}

\item{tripsTable}{tibble of output_trips (from read_output_trips())}

\item{shapeTable}{sf object(data.frame with geometries), can be created using st_read()}
}
\value{
Bar chart plot of the distance traveled by trip type
}
\description{
Takes the data frame output_trips (from \link{read_output_trips()})
to plot a bar chart of the traveled distance of each trip type (originating, destinating, inside, outside)
as identified with a shape file of the project area.
}
