% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{plot_distance_by_spatialcat_barchart}
\alias{plot_distance_by_spatialcat_barchart}
\title{Bar Chart with tripType on x-axis and travelled distance on y-axis}
\usage{
plot_distance_by_spatialcat_barchart(
  trips_table,
  shape_table,
  crs,
  euclidean = FALSE
)
}
\arguments{
\item{trips_table}{tibble of trips_output (from \link{read_output_trips})}

\item{crs}{numeric of EPSG code or proj4string, can be found in network file from output directory of MATSim simulation}

\item{euclidean}{Logical value indicating whether to calculate the average as euclidean distance or as travel distance. Default is FALSE, which calculates the average traveled distance.}

\item{shape_Table}{sf object(data frame with geometries), can be created using st_read(), is used to categorize the trips.}
}
\value{
Bar Chart plot of distance traveled by spatial type
}
\description{
Takes Table trips_output (from \link{read_output_trips}),
to plot bar chart with with values that represent
travelled distance of each tripType related to the shapeTable
}
