% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{plot_distcat_by_mainmode_barchart}
\alias{plot_distcat_by_mainmode_barchart}
\title{Bar Chart with distance traveled on x-axis and number of trips on y-axis}
\usage{
plot_distcat_by_mainmode_barchart(
  trips_table,
  unite.modes = character(0),
  united.name = "united",
  dist_column = "dist_cat",
  distances_array = c(1000, 2000, 5000, 10000, 20000, 50000, 1e+05)
)
}
\arguments{
\item{trips_table}{tibble of trips_output (from readTripsTable())}

\item{unite.modes}{vector of character strings,
changes names of chosen modes in the column main_mode to a new chosen name (i.e. drtNorth and drtSouth to drt),
using the function (\link{process_rename_mainmodes})}

\item{united.name}{character string, specifies the name of the united mode}
}
\value{
Bar Chart plot of count of trips among distance travelled
}
\description{
Takes the data frame trips_output (from \link{readTripsTable()}) and categorizes the traveled distances into pre-defined bins
to plot a histogram of the traveled distances. (Bins: 1000,2000,5000,10000,20000,50000,100000 (m))
Using the parameter unite.modes, specific modes can be renamed into one with the name specified with united.name (by default 'united')
}
