% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{plot_mainmode_piechart}
\alias{plot_mainmode_piechart}
\title{Plot the distribution of modes as a pie chart}
\usage{
plot_mainmode_piechart(
  trips_table,
  unite.modes = character(0),
  united.name = "united",
  dump.output.to = matsimDumpOutputDirectory
)
}
\arguments{
\item{trips_table}{tibble of output_trips (from \link{read_output_trips})}

\item{unite.modes}{vector of character strings,
changes names of chosen modes in the column main_mode to a new chosen name (i.e. drtNorth and drtSouth to drt),
using the function (\link{process_rename_mainmodes})}

\item{united.name}{character string, specifies the name of the united mode}
}
\value{
pie chart plot of transport mode distribution, values given in percent
}
\description{
Uses the data frame output_trips (from \link{read_output_trips}),
to plot a pie chart of the modal split using the column main_mode
}
\details{
The function automatically detects the modes plots a pie chart.
Using the parameter unite.modes, specific modes can be renamed into one with the name specified with united.name (by default 'united')
}
