% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{plot_map_trips_by_spatialcat}
\alias{plot_map_trips_by_spatialcat}
\title{Plots all trips categorized by trip type (inside, outside, origin and destinating) on a map}
\usage{
plot_map_trips_by_spatialcat(trips_table, shape_table, crs, optimized = FALSE)
}
\arguments{
\item{crs}{numeric representation of the EPSG code or proj4string for the corresponding coordinate system
of the trip coordinates, can be found in network file from output directory of MATSim simulation}

\item{optimized}{bool, by default FALSE and gives interactive plot using leaflet, if TRUE creates image with ggplot}

\item{table}{tibble of output_trips (from read_output_trips())}

\item{shapeTable}{sf object(data.frame with geometries), can be created using st_read()}
}
\value{
plot that contains every trip with defined trip type
}
\description{
Plots all trips categorized by trip type (inside, outside, origin and destinating) on a map
}
