% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{plot_map_trips_by_spatialcat}
\alias{plot_map_trips_by_spatialcat}
\title{Plots every type of trips(inside, outside, origin and destinating) on map}
\usage{
plot_map_trips_by_spatialcat(trips_table, shape_table, crs, optimized = FALSE)
}
\arguments{
\item{trips_table}{tibble of trips_output (from \link{read_output_trips()})}

\item{shape_table}{sf object(data.frame with geometries), can be received by using st_read(path_to_geographical_file)}

\item{crs}{numeric of EPSG code or proj4string, can be found in network file from output directory of MATSim simulation}

\item{optimized}{bool, by default FALSE and gives interactive plot using leaflet, if TRUE using image with ggplot}
}
\value{
plot that contains every trip with defined trip type
}
\description{
Plots every type of trips(inside, outside, origin and destinating) on map
}
