% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{plot_spatialtype_by_shape_piechart}
\alias{plot_spatialtype_by_shape_piechart}
\title{Plots distribution of every type of trips(inside, outside, origin and destinating) in Pie Chart
XXXX}
\usage{
plot_spatialtype_by_shape_piechart(trips_table, shape_table, crs)
}
\arguments{
\item{trips_table}{tibble of trips_output (from \link{read_output_trips})}

\item{shape_table}{sf object(data.frame with geometries), can be received by using st_read(path_to_geographical_file)}

\item{crs}{numeric of EPSG code or proj4string, can be found in network file from output directory of MATSim simulation}
}
\value{
plot with percentage of each type of trips
}
\description{
Plots distribution of every type of trips(inside, outside, origin and destinating) in Pie Chart
XXXX
}
