% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{plot_trips_count_by_deptime_and_mainmode_linechart}
\alias{plot_trips_count_by_deptime_and_mainmode_linechart}
\title{Line plot with departure time x-axis and number of trips on y-axis}
\usage{
plot_trips_count_by_deptime_and_mainmode_linechart(
  trips_table,
  unite.columns = character(0),
  united.name = "united"
)
}
\arguments{
\item{unite.columns}{vector of character strings, that represent patterns of columns to be united, changes name of all transport modes in the tibble copy to united.name = "united" that matches PATTERNS given in unite.columns}

\item{united.name}{character string, if columns were united, you can specify name for the resulting column in chart}

\item{tripsTable}{tible of trips_output (from readTripsTable())}
}
\value{
Line Chart plot of trips count by departure mode per mode
}
\description{
Takes Table trips_output (from readTripsTable()),
to make line plot with with values that represent
count of trips for a specific departure time by main_mode
Using parameters unite.columns, specific columns could be given, to unite them in 1 mode with the name united.name(by default 'united')
}
