% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simwrapper.R
\name{prepareComparisonSimwrapperDashboardFromTable}
\alias{prepareComparisonSimwrapperDashboardFromTable}
\title{Creates comparison dashboard for the two given tables or folders with data}
\usage{
prepareComparisonSimwrapperDashboardFromTable(
  table1,
  table2,
  dump.output.to = matsimDumpOutputDirectory,
  append = FALSE
)
}
\arguments{
\item{table1}{trips_output tibble from readTripsTable()}

\item{table2}{trips_output tibble from readTripsTable()}

\item{dump.output.to}{output folder, yaml files and plots from the following functions are saved there:
plotModalSplitBarChart(),plotModalSplitPieChart(),plotModalShift().}

\item{append}{specifies if the output folder should be erased before creating, default = FALSE}
}
\value{
generates folder with content for simwrapper from a trips table
}
\description{
Creates comparison dashboard for the two given tables or folders with data
}
