% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripsOutput.R
\name{prepareSimwrapperDashboardFromFolder}
\alias{prepareSimwrapperDashboardFromFolder}
\title{Creates dashboard for the given table or folder with data}
\usage{
prepareSimwrapperDashboardFromFolder(folder, append = FALSE)
}
\arguments{
\item{folder}{specifies data source folder with tripsOutput}

\item{append}{specifies if the ouput folder should be erased before creating}
}
\value{
plot that contains every trip with defined trip type
}
\description{
Creates dashboard for the given table or folder with data
}
