% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simwrapper.R
\name{prepareSimwrapperDashboardFromFolder}
\alias{prepareSimwrapperDashboardFromFolder}
\title{Creates dashboard for the given folder with data}
\usage{
prepareSimwrapperDashboardFromFolder(
  folder,
  dump.output.to = matsimDumpOutputDirectory,
  append = FALSE
)
}
\arguments{
\item{folder}{file path, specifies data source folder that includes a output_trips.csv file}

\item{dump.output.to}{output folder, yaml files are saved there}

\item{append}{specifies if the ouput folder should be erased before creating, default = FALSE}
}
\value{
tibble of output_trips from folder and content needed for Simwrapper
}
\description{
Creates dashboard for the given folder with data
}
