% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripsOutput.R
\name{prepareSimwrapperDashboardFromFolder}
\alias{prepareSimwrapperDashboardFromFolder}
\title{Creates dashboard for the given table or folder with data}
\usage{
prepareSimwrapperDashboardFromFolder(folder, append = FALSE)
}
\arguments{
\item{folder}{specifies data source folder with tripsOutput}

\item{append}{specifies if the ouput folder should be erased before creating}
}
\value{
tibble of output_trips from folder. Generates content needed for Simwrapper
}
\description{
Creates dashboard for the given table or folder with data
}
