% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simwrapper.R
\name{prepareSimwrapperDashboardFromTable}
\alias{prepareSimwrapperDashboardFromTable}
\title{Creates dashboard for the given table or folder with data}
\usage{
prepareSimwrapperDashboardFromTable(
  table,
  dump.output.to = matsimDumpOutputDirectory,
  append = FALSE
)
}
\arguments{
\item{table}{trips_output tibble from readTripsTable()}

\item{dump.output.to}{folder that saves and configures yaml for simwrapper dashboard and all plots using functions:
plotModalSplitBarChart(),plotModalSplitPieChart(),plotModalShift().}

\item{append}{specifies if the ouput folder should be erased before creating}
}
\value{
generates folder with content for simwrapper out of trips table
}
\description{
Creates dashboard for the given table or folder with data
}
