% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{process_append_distcat}
\alias{process_append_distcat}
\title{Adds additional \strong{dist_cat} column based on the \strong{distances_array} parameter}
\usage{
process_append_distcat(
  trips_table,
  distances_array = c(1000, 2000, 5000, 10000, 20000, 50000, 1e+05)
)
}
\arguments{
\item{trips_table}{tibble of output_trips (from \link{read_output_trips})}

\item{distances_array}{numeric vector, represents segments for distance categories ordered (in meters),
standard value: c(1000,2000,5000,10000,20000,50000,100000)}
}
\value{
trips table with added the added column dist_cat
}
\description{
Categorizes each trip into a distance bin based on traveled distance.
E.g. distance traveled is 1500, distances array is (1000,2000), then the category is "1000-2000".
}
