% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{process_append_spatialcat}
\alias{process_append_spatialcat}
\title{Appends an additional column with the trip type to output_trips}
\usage{
process_append_spatialcat(trips_table, shape_table, crs)
}
\arguments{
\item{trips_table}{tibble of output_trips (from read_output_trips())}

\item{crs}{numeric representation of the EPSG code or proj4string for the corresponding coordinate system
of the trip coordinates, can be found in network file from output directory of MATSim simulation}

\item{shapeTable}{sf object(data.frame with geometries), can be created using st_read()}
}
\value{
tibble, with additional column containing the trip type
}
\description{
Using a shape_file of the project area an additional column is created, categorizing all trips into the following categories.
inside: trips that start and end inside the given shape
originating: trips that start inside the shape and end outside of the shape
destinating: trips that end inside the shape and start outside of the shape
outside: trips that start and end outside of the shape
}
\details{
XXXX Can this be deleted? Is this relevant here?
Please be aware that this \link{process_filter_by_shape} currently only works, when one geometry is loaded.)
}
