% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{process_filter_by_shape}
\alias{process_filter_by_shape}
\title{Filtering of trips_table(from readTripsTable) depending on how they located in given shape}
\usage{
process_filter_by_shape(trips_table, shape_table, crs, spatial_type = "inside")
}
\arguments{
\item{trips_table}{tibble of trips_output (from readTripsTable())}

\item{crs}{numeric of EPSG code or proj4string, can be found in network file from output directory of MATSim simulation}

\item{spatial_type}{character parameter,categories of spatial information used to specify which kind of trips should be filtered. Takes as input 4 templates: inside, outside, originating, destinating}

\item{shapeTable}{sf object(data.frame with geometries), can be received by using st_read(path_to_geographical_file)}
}
\value{
tibble, with filtered trips depending on shapeTable and special flags (see Description)
}
\description{
Takes trips_table and shapeTable(sf object from file representing geographical data, can be received by using function st_read(path_to_file).
Please be aware that this filterByRegion currently only works, when one geometry is loaded.)
transforms both objects to match mutual CRS(network.xml from MATSimOutputDirectory)
and filters the trips from table depending on *.inshape flags:
if start.inshape = TRUE & end.inshape = TRUE return table that contains trips \strong{inside} of the given shape
if start.inshape = TRUE & end.inshape = FALSE return table that contains trips which \strong{originating} in the shape
if start.inshape = FALSE & end.inshape = TRUE return table that contains trips which \strong{destinating} in the shape
if start.inshape = FALSE & end.inshape = FALSE return table that contains trips which \strong{outside} of the given shape
}
