% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{process_get_travdistance_distribution}
\alias{process_get_travdistance_distribution}
\title{Calculates the average distance for each mode}
\usage{
process_get_travdistance_distribution(trips_table, euclidean = FALSE)
}
\arguments{
\item{trips_table}{tibble of output_trips (from read_output_trips())}

\item{euclidean}{boolean, specifies if euclidean or traveled distance should be used}
}
\value{
a tibble of main modes and their average distance
}
\description{
Calculates the average distance for each mode
}
