% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{process_get_travelwaittime_by_mainmode}
\alias{process_get_travelwaittime_by_mainmode}
\title{Calculates the average travel and wait time
XXXX minute_ but hourS and secondS?}
\usage{
process_get_travelwaittime_by_mainmode(trips_table, time_format = "minute")
}
\arguments{
\item{trips_table}{tibble of output_trips (from read_output_trips())}

\item{time_format}{standard value is "minute", can also be "hours" or "seconds"}
}
\value{
tibble with main modes and their average wait and travel time
}
\description{
Calculates the average travel and wait time
XXXX minute_ but hourS and secondS?
}
