% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{process_rename_category}
\alias{process_rename_category}
\title{Replace two or more named categories of any column with a new joint category name}
\usage{
process_rename_category(
  trips_table,
  unite_template = character(0),
  united_name = "united",
  column = "main_mode"
)
}
\arguments{
\item{trips_table}{tibble of output_trips (from \link{\code{read_output_trips}})}

\item{unite_template}{vector of character strings,
changes names of chosen categories to a new specified name (i.e. drtNorth and drtSouth to drt)}

\item{united_name}{character string, specifies the name of the united categories}

\item{column}{tibble column name, which will be used in occurrences renaming}
}
\value{
modified trips table tibble
}
\description{
The function process_rename_category takes a tibble
as input and replaces all occurrences of the specified categories in a column
with another name provided in united_name.
}
