% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{process_rename_mainmodes}
\alias{process_rename_mainmodes}
\title{Replace two or more modes in the column main_mode with a new joint mode}
\usage{
process_rename_mainmodes(
  trips_table,
  unite_modes = character(0),
  united_name = "united"
)
}
\arguments{
\item{trips_table}{tibble of output_trips (from \link{\code{read_output_trips}})}

\item{unite_modes}{vector of character strings,
changes names of chosen modes in the column main_mode to a new chosen name (i.e. drtNorth and drtSouth to drt)}

\item{united_name}{character string, specifies the name of the united mode}
}
\value{
tibble with new main_mode representation
}
\description{
The function process_rename_mainmodes takes the tibble trips_table
as input and replaces all occurrences of the specified modes in column main_mode
with another name provided in united_name.
}
