% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{process_rename_mainmodes}
\alias{process_rename_mainmodes}
\title{XXXX}
\usage{
process_rename_mainmodes(
  trips_table,
  unite.modes = character(0),
  united.name = "united"
)
}
\description{
XXXX
}
