% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{processDtvEstimationQuality}
\alias{processDtvEstimationQuality}
\title{Categorize DTV deviation and aggregate data}
\usage{
processDtvEstimationQuality(
  joinedFrame,
  aggr = TRUE,
  ll = ~x * 0.8 - 200,
  ul = ~x * 1.2 + 200
)
}
\arguments{
\item{joinedFrame}{A tibble from mergeCountsAndLinks}

\item{aggr}{Boolean, if categorized data should returned aggregated, default is TRUE.}

\item{ll}{Formula to calculate lower limit of the quality label 'exact', default = 0.8*x - 200}

\item{ul}{Formula to calculate lower limit of the quality label 'exact', default = 1.2*x + 200}
}
\value{
A long-format tibble, which contains share of estimation quality for each scenario and link type, if aggr is FALSE disaggregated data is returned
}
\description{
Takes a tibble from mergeCountsAndLinks.
Deviation between count volumes and Linkstats is calculated and
categorized (i.e. deviation of 1.2 means 20 percent more DTV in MATSim than in counts).
If parameter 'aggr' is set to TRUE, data will be aggregated for each run and link type.
Can be used to visualize model quality by link type and to compare several runs.
}
\details{
Estimation quality is determined by the 'cut' function, limits for the label
'exact' can be adjusted by tuning the parameters 'll' (lower limit) and 'ul' (upper limit)
}
