% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{processLinkStatsDtvDistribution}
\alias{processLinkStatsDtvDistribution}
\title{Categorize DTV and calculate DTV distribution}
\usage{
processLinkStatsDtvDistribution(joinedFrame)
}
\arguments{
\item{joinedFrame}{A tibble from mergeCountsAndLinks}
}
\value{
A long-format tibble which contains share of DTV Categories
}
\description{
Takes a tibble from mergeCountsAndLinks. DTV is categorized into bins. Finally
data is aggregated to calculate DTV distribution in each link type category,
excluding 'residential' and 'unclassified'
Data can be used to create multiple geom_col plots to visualize and compare
DTV distributions between count data and several MATSim runs
}
