% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{processLinkStatsDtvDistribution}
\alias{processLinkStatsDtvDistribution}
\title{Categorize daily traffic volume (DTV) and calculate DTV for different link types.}
\usage{
processLinkStatsDtvDistribution(joinedFrame, from = 0, to = 40000, by = 5000)
}
\arguments{
\item{joinedFrame}{A tibble from mergeCountsAndLinks}

\item{from}{Integer. Lower limit for count bin, default = 0.}

\item{to}{Integer. Upper limit for count bins, default = 40000.}

\item{by}{Integer. Size of each count bin, default = 5000.}
}
\value{
A long-format tibble which contains share of DTV for link types
}
\description{
Takes a tibble from mergeCountsAndLinks. DTV is categorized into bins. Finally
data is aggregated to calculate DTV distribution in each link type category,
excluding 'residential' and 'unclassified'
Data can be used to create multiple geom_col plots to visualize and compare
DTV distributions between count data and several MATSim runs
}
