% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trips.R
\name{read_output_trips}
\alias{read_output_trips}
\title{Load MATSim output_trips table into memory}
\usage{
read_output_trips(input_path = ".", n_max = Inf)
}
\arguments{
\item{input_path}{character string, path to the MATSim output directory or http link to the file}

\item{n_max}{optional, integer, maximum number of lines to read, standard value is Inf}
}
\value{
tibble of output_trips
}
\description{
Loads MATSim output_trips.csv from file or archive creating a tibble
}
