% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{readLinkStats}
\alias{readLinkStats}
\title{Load linkstats as tibble into memory}
\usage{
readLinkStats(runId, file, sampleSize = 0.25)
}
\arguments{
\item{runId}{Id to tag columns with DTV}

\item{file}{File to load. Must be an .csv or .tsv file with comma separator}

\item{sampleSize}{sample size of the MATSim scenario to scale DTV values}
}
\value{
Tibble with link stats for each qsim mode
}
\description{
Reads Linkstats as .tsv created from LinkStats.class
as dataframe into memory.
Counts can be provided in any time bins.
Counts can be provided for any qsim mode. The argument networkModes is used to
select and filter the columns.
}
