% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripsOutput.R
\name{readTripsTable}
\alias{readTripsTable}
\title{Works and analyses with trips_output file from MatSim
Starting point of the workflow is readTripsTable
Plotting is based on main_mode attr of trips_output}
\usage{
readTripsTable(pathToMATSimOutputDirectory = ".")
}
\description{
transformToSf changes representation of trips_output to geographical with geometry
filterByRegion takes location of shapeFile and filters the trips_output excluding trips out of shapeFile
}
