% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripsOutput.R
\name{readTripsTable}
\alias{readTripsTable}
\title{Load MATSIM output_trips table into Memory}
\usage{
readTripsTable(input_path = ".")
}
\arguments{
\item{input_path}{character string, path to matsim output directory or http link to the file.}
}
\value{
tibble of trips_output
}
\description{
Loads a MATSim CSV output_trips from file or archive,
creating a tibble with columns as in csv file
}
