% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripsOutput.R
\name{transformToSf}
\alias{transformToSf}
\title{Transforms trips_table tibble (from readTripsTable) from tibble to sf (table with attribute features and geometry feature)}
\usage{
transformToSf(table, crs, geometry.type = st_multipoint())
}
\arguments{
\item{table}{tibble of trips_output (from readTripsTable())}

\item{crs}{numeric of EPSG code or proj4string, can be found in network file from output directory of MATSim simulation}

\item{geometry.type}{function of sf transformation, geometry.type can be (by default is st_multipoint())
!!!st_point()-resulting table contains 2 geometries start_wkt and end_wkt, representing start and end POINTs, and have type POINT!!!  or
!!!st_multipoint()-resulting table contains 1 geometry wkt, representing start and end POINTS as MULTIPOINT!!! or
!!!st_linestring() - resulting table contains 1 geometry wkt, representing line between start and end points as LINESTRING!!!}
}
\value{
sf object (data.frame with geometries depending to geometry.type)
}
\description{
Takes trips_table (from readTripsTable) and transforms trips_table to sf object using start_x, end_x, start_y, end_y as a geometry features
deletes from resulting data.frame start_x, end_x, start_y, end_y.
And adds wkt column, if geometry.type = st_mulitpoint(), or geometry.type = st_linestring()
Or adds start_wkt and end_wkt, if geometry.type = st_point()
Added column/columns projected to given CRS (coordinate reference system),
that can be taken from network file of MATSimOutputDirectory
}
\details{
Function also sets attribute geometry.type to resulting table to character value of "POINT","MULTIPOINT","LINESTRING"
to get which type of table was generated, if it is needed
}
